+++ rosetta.go
@@ -26,7 +26,7 @@
 package rosetta
 
 // Version is the semantic version of this SDK release.
-const Version = "0.3.0"
+const Version = "0.3.1"
 
 // Protocol identifies one of the supported wire protocols.
 type Protocol string